Changed all functions in vm/core/src/memory.c taking a memory size as jint to use size_t instead. rvmAllocateMemoryForArray() will now detect overflow and throw OOME if the total size of the requested array is larger than the maximum size_t value.

+++ vm/core/src/test/test_call0.c
@@ -18,7 +18,7 @@
 #include "../private.h"
 #include "CuTest.h"
 
-void* rvmAllocateMemory(Env* env, int size) {
+void* rvmAllocateMemory(Env* env, size_t size) {
     return calloc(1, size);
 }
